backend/requirements.txt
requests>=2.28.0 # HTTP library for API calls
Django>=4.2,<5.0
djangorestframework>=3.14,<3.15
psycopg2-binary==2.9.9
setuptools>=65.0.0 # Required for pkg_resources
djangorestframework-simplejwt>=5.2,<5.3
openai>=1.0.0
django-cors-headers>=3.0.0 # For CORS
whitenoise[brotli]>=6.0.0 # For static files, with brotli support
django-filter>=23.0 # For filtering in DRF
dj-database-url>=2.0.0 # For parsing DATABASE_URL environment variable
gunicorn>=21.2.0 # For production server
python-dotenv>=1.0.0 # For loading environment variables from .env files
ics>=0.7.2 # For generating iCalendar files
pytz>=2023.3 # For timezone handling
# Example of how you might pin a specific version:
# openai==1.3.0
